refactor: Use a nested class for ProfessionItemFilterMatcher
