+++ common/src/main/java/com/wynntils/services/itemfilter/filters/ProfessionItemFilter.java
@@ -4,6 +4,10 @@
  */
 package com.wynntils.services.itemfilter.filters;
 
+import com.wynntils.models.items.WynnItem;
+import com.wynntils.models.items.items.game.GatheringToolItem;
+import com.wynntils.models.items.items.game.IngredientItem;
+import com.wynntils.models.items.items.game.MaterialItem;
 import com.wynntils.models.profession.type.ProfessionType;
 import com.wynntils.services.itemfilter.type.ItemFilter;
 import com.wynntils.services.itemfilter.type.ItemFilterMatcher;
@@ -22,6 +26,34 @@ public class ProfessionItemFilter extends ItemFilter {
             return ErrorOr.error(getTranslation("invalidProfession", inputString));
         } else {
             return ErrorOr.of(new ProfessionItemFilterMatcher(profession));
+        }
+    }
+
+    private static class ProfessionItemFilterMatcher implements ItemFilterMatcher {
+        private final ProfessionType profession;
+
+        protected ProfessionItemFilterMatcher(ProfessionType profession) {
+            this.profession = profession;
+        }
+
+        @Override
+        public boolean matches(WynnItem wynnItem) {
+            if (profession == null) {
+                return false;
+            } else if (wynnItem instanceof IngredientItem ingredient) {
+                return ingredient.getIngredientInfo().professions().contains(profession);
+            } else if (wynnItem instanceof MaterialItem materialItem) {
+                return materialItem
+                        .getMaterialProfile()
+                        .getResourceType()
+                        .getMaterialType()
+                        .getProfessionType()
+                        == profession;
+            } else if (wynnItem instanceof GatheringToolItem gatheringToolItem) {
+                return gatheringToolItem.getToolProfile().toolType().getProfessionType() == profession;
+            }
+
+            return false;
         }
     }
 }
